requirements.txt
jsonformer>=0.12.0
langchain>=0.1.1
langchain-community
pandas>=1.5.3
wikipedia>=1.4.0
loguru>=0.7.2
datasets>=2.15.0
jsonlines>=4.0.0
streamlit>=1.31.0
requests>=2.25.0
google-generativeai
openai>=1.0.0
# ollama>=0.1.0
recbole
torch==2.5.0 --index-url https://download.pytorch.org/whl/cu124
scipy==1.12.0
numpy==1.24.0